requirements.txt
numpy >= 1.20.1
pygame >= 2.1.1
pyinstaller >= 5.1
Pillow >= 9.1.1
numba >= 0.55.2